+++ lib/google-api/client.rb
@@ -16,12 +16,12 @@ module GoogleAPI
     # Build an AccessToken object from OAuth2. Check if the access token is expired, and if so,
     # refresh it and save the new access token returned from Google.
     def access_token
-      @access_token = ::OAuth2::AccessToken.new(client, object.oauth_access_token,
-        refresh_token: object.oauth_refresh_token,
-        expires_at: object.oauth_access_token_expires_at.to_i
+      @access_token = ::OAuth2::AccessToken.new(client, object.oauth_hash[:access_token],
+        refresh_token: object.oauth_hash[:refresh_token],
+        expires_at: object.oauth_hash[:expires_at].to_i
       )
       if @access_token.expired?
-        GoogleAPI.logger.info "Access Token expired for #{object.class.name}(#{object.id}), refreshing..." if defined?(::Rails)
+        GoogleAPI.logger.info "Access Token expired for #{object.class.name}(#{object.id}), refreshing..."
         @access_token = @access_token.refresh!
         object.update_access_token!(@access_token.token)
       end
@@ -38,16 +38,25 @@ module GoogleAPI
       )
     end
 
-    # Each API that Google offers us is instantiated within its own method below.
-    # If you want to find all calendars for a user, you would do this:
+    # We build the appropriate API here based on the method name passed to the Client.
+    # For example:
     #
-    #   GoogleAPI::Client.new(User.first).calendar.all
+    #   User.find(1).google.drive
     #
-    # See GoogleAPI::Calendar for more information about the Calendar API.
-    %w(calendar drive).each do |api|
-      define_method api do
-        "GoogleAPI::#{api.capitalize}".constantize.new(access_token)
+    # We will then discover and cache the Google Drive API for future use.
+    # Any methods chained to the resultant API will then be passed along to the
+    # instantiaed class. Read the documentation for GoogleAPI::API#method_missing for
+    # more information.
+    def method_missing(api, *args)
+      unless GoogleAPI.discovered_apis.has_key?(api)
+        GoogleAPI.logger.info "Discovering the #{api} Google API..."
+        response = access_token.get("https://www.googleapis.com/discovery/v1/apis?preferred=true&name=#{api}").parsed['items']
+        super unless response # Raise a NoMethodError if Google's Discovery API does not return a good response
+        discovery_url = response.first['discoveryRestUrl']
+        GoogleAPI.discovered_apis[api] = access_token.get(discovery_url).parsed
       end
+
+      API.new(access_token, api, GoogleAPI.discovered_apis[api]['resources'])
     end
 
   end