+++ spec/google-api/api/drive/permission_spec.rb
@@ -1,5 +1,0 @@
-require 'spec_helper'
-
-describe GoogleAPI::Drive::Permission do
-
-end