+++ spec/spec_helper.rb
@@ -6,6 +6,12 @@ end
 require 'rspec'
 require 'google-api'
 
+Faraday.default_adapter = :test
+
 # Requires supporting files with custom matchers and macros, etc,
 # in ./support/ and its subdirectories.
 Dir["#{File.dirname(__FILE__)}/support/**/*.rb"].each {|f| require f}
+
+RSpec.configure do |config|
+  config.after(:each) { GoogleAPI.reset_environment! }
+end