+++ spec/support/oauth2_response.rb
@@ -1,0 +1,17 @@
+class OAuth2Response
+
+  attr_accessor :status
+
+  def initialize(status_code = 200)
+    @status = status_code
+  end
+
+  def headers
+    {}
+  end
+
+  def parsed
+    fixture(:drive_files)
+  end
+
+end