+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 ## v0.0.1.beta (TBD)
 
+* Full test suite for all classes and modules.
+* Rethink and rebuild the API integration. Now discover the API and build methods dynamically! Works with the following Google APIs:
+  * Calendar
+  * Drive
 * Add logger configuration.
 * Add the #google method to an oauthable object.
-* Refactored some settings into class methods for each API class.
-* Built out more of the Drive API, including the Permissions sub-API.
 * Add the #patch method to OAuth2.
 
 ## v0.0.1.alpha (2012-08-21)