+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## v0.0.1.beta (TBD)
+## v0.0.1.beta (2012-09-05)
 
 * Full test suite for all classes and modules.
 * Rethink and rebuild the API integration. Now discover the API and build methods dynamically! Works with the following Google APIs: